api/requirements.txt
azure-functions
requests
azure-functions
azure-data-tables